requirements/base.txt
Django>=1.9
django-environ>=0.4.0
django-braces>=1.8.1
django-crispy-forms>=1.5.2
django-admin-bootstrapped>=2.5.6
django-authtools>=1.4.0
django-oauth-toolkit>=0.12.0
easy-thumbnails>=2.3
djangorestframework==3.6.2
markdown==2.6.8
django-filter==1.0.2
pint==0.8
django-model-utils==3.0.0
django-positions==0.5.4